# pkg/swagger/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Chat Server API",
    "description": "Chat Server service",
    "version": "1.0",
    "contact": {
      "name": "Elmir Rasulov",
      "email": "elmir.rasulov.2002@mail.ru"
    }
  },
  "tags": [
    {
      "name": "chatServer"
    }
  ],
  "host": "localhost:8082",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create": {
      "post": {
        "summary": "Создания чата",
        "operationId": "chatServer_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/chatServerCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/chatServerCreateRequest"
            }
          }
        ],
        "tags": [
          "chatServer"
        ]
      }
    },
    "/v1/delete": {
      "get": {
        "summary": "Удаления чата",
        "operationId": "chatServer_Delete",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "chatServer"
        ]
      }
    },
    "/v1/send-messages": {
      "post": {
        "summary": "Отправка сообщения",
        "operationId": "chatServer_SendMessage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/chatServerSendMessageRequest"
            }
          }
        ],
        "tags": [
          "chatServer"
        ]
      }
    }
  },
  "definitions": {
    "chatServerCreateRequest": {
      "type": "object",
      "properties": {
        "usernames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "chatServerCreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "chatServerSendMessageRequest": {
      "type": "object",
      "properties": {
        "chatId": {
          "type": "string",
          "format": "int64"
        },
        "from": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}